Major Changes
Restructured packages and reorganized classes
Added Builder to each of the node classes
Set the MethodNode to include Statements in its JSON representation
Updated each of the Deserializers
Modified each Node to be more restrictive
Added a highest level of abstraction interface, INode, to the node
hierarchy
Added an abstract class, AbstractNode, which is the base class of all
Nodes.
Added a new class under AbstractNode and sibling to CodeNode called
StructureNode.
Reorganized the Node hierarchy to be more reprentative of a software
system.
Documented all classes
Added new tests.
Added NamespaceNode to handle namespaces
Added Relationship to start handling relationships between classes which
are not accounted for by the implicit tree relationships.


Signed-off-by: Isaac Griffith <isaacgriffith@gmail.com>